ch(setup): Restructure File Org - Separate Concerns
